+++ src/api/consumer/lib/get-component-logs.js
@@ -4,5 +4,5 @@ import { BitId } from '../../../bit-id';
 
 export default function getComponentLogs(id: string) {
   return loadConsumer()
-    .then(consumer => consumer.scope.loadComponentLogs(BitId.parse(id, consumer.scope.name)));
+    .then(consumer => consumer.scope.loadComponentLogs(BitId.parse(id)));
 }