+++ src/bit-id/bit-id.js
@@ -5,7 +5,6 @@ import { InvalidBitId, InvalidIdChunk } from './exceptions';
 import {
   LATEST_BIT_VERSION,
   VERSION_DELIMITER,
-  LOCAL_SCOPE_NOTATION,
   NO_PLUGIN_TYPE,
 } from '../constants';
 import { isValidIdChunk, isValidScopeName } from '../utils';
@@ -78,7 +77,7 @@ export default class BitId {
     return path.join(this.box, this.name, this.scope, this.version);
   }
 
-  static parse(id: ?string, realScopeName: ?string, version: string = LATEST_BIT_VERSION): ?BitId {
+  static parse(id: ?string, version: string = LATEST_BIT_VERSION): ?BitId {
     if (!id || id === NO_PLUGIN_TYPE) { return null; }
     if (id.includes(VERSION_DELIMITER)) {
       const [newId, newVersion] = id.split(VERSION_DELIMITER);
@@ -89,17 +88,13 @@ export default class BitId {
     const idSplit = id.split('/');
     if (idSplit.length === 3) {
       const [scope, box, name] = idSplit;
-      if (scope === LOCAL_SCOPE_NOTATION && !realScopeName) {
-        throw new Error('real scope name is required in bitId.parse with @this notation');
-      }
-      const digestScopeName = scope === LOCAL_SCOPE_NOTATION ? realScopeName : scope;
-      if (!isValidIdChunk(name) || !isValidIdChunk(box) || !isValidScopeName(digestScopeName)) {
+      if (!isValidIdChunk(name) || !isValidIdChunk(box) || !isValidScopeName(scope)) {
         // $FlowFixMe
-        throw new InvalidIdChunk(`${digestScopeName}/${box}/${name}`);
+        throw new InvalidIdChunk(`${scope}/${box}/${name}`);
       }
       // $FlowFixMe (in this case the realScopeName is not null)
       return new BitId({
-        scope: digestScopeName,
+        scope,
         box,
         name,
         version