+++ src/bit-id/bit-ids.js
@@ -36,7 +36,7 @@ export default class BitIds extends Array<BitId> {
     const array = [];
 
     forEach(dependencies, (version, id) => {
-      array.push(BitId.parse(id, null, version));
+      array.push(BitId.parse(id, version));
     });
 
     return new BitIds(...array);