+++ CHANGELOG.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- windows support
+- fix Error: "Cannot find module './build/Release/DTraceProviderBindings'" when installing via Yarn
+- fix the local and remote search
+- support deleting files from a component
 - fix the internal ci-update command where an environment has a tester without a compiler 
 - improved commit, add, export and status outputs
 - improved bit help
+- fix bit config command for linux
 
 ## [0.10.5] - 2017-08-16
 - improved commit, add, export and status outputs